data-etl/pyproject.toml
[tool.poetry]
name = "parking-jawn-data"
version = "0.1.0"
description = "Toolkit to upload data files for Parking Jawn to AWS"
authors = ["Nick Hand <nicholas.adam.hand@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.20.49"
pandas = "^1.4.0"
loguru = "^0.6.0"
python-dotenv = "^0.19.2"
carto2gpd = ">1,<2"
click = "^8.0.3"
Fiona = "1.8.21"
Rtree = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
isort = "^5.10.1"
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
parking-jawn-data = "parking_jawn_data.__main__:main"